kavachai/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
pynacl==1.5.0
lark==1.2.2
redis[hiredis]==5.1.0
litellm==1.50.0
mcp==1.1.0
aiosqlite==0.20.0
httpx==0.27.0
python-multipart==0.0.9
